fix(openui): render dev playground client-only (browser-only chart/map deps)
